Add backup featue
Adds backupre feature which has been tested for partially on MacOs. Remaining testing remains to see if the full functionality works on Linux. There is an automatic backup of the database which runs every minute, as well as backups of some dummy system configurations which will be saved to a table in a databse. Routes have been created to backup as well as to save configurations